LayoutTests/fast/forms/time/ValidityState-typeMismatch-time.html

<!DOCTYPE html>
<html>
<head>
<script src="../../js/resources/js-test-pre.js"></script>
</head>
<body>
<p id="description"></p>
<div id="console"></div>
<script>
description('This test aims to check for typeMismatch flag with type=time input fields');
var i = document.createElement('input');
i.type = 'time';

function check(value, mismatchExpected, disabled)
{
    i.value = value;
    i.disabled = !!disabled;
    var actual = i.validity.typeMismatch;
    var didPass = actual == mismatchExpected;
    var resultText = '"' + value + '" is ' + (didPass ? 'a correct ' : 'an incorrect ') + (actual ? 'invalid' : 'valid') + ' time string' + (disabled ? ' when disabled.' : '.');
    if (didPass)
        testPassed(resultText);
    else
        testFailed(resultText);
}

function shouldBeValid(value, disabled)
{
    check(value, false, disabled);
}

function shouldBeInvalid(value)
{
    check(value, true);
}

// Valid values
shouldBeValid('');
shouldBeValid('00:00');
shouldBeValid('23:59');
shouldBeValid('23:59:59');
shouldBeValid('23:59:59.1');
shouldBeValid('23:59:59.12');
shouldBeValid('23:59:59.123');
shouldBeValid('23:59:59.1234567890');
shouldBeValid('00:00:00.0000000000');

// Invalid values
shouldBeInvalid(' 00:00 ');
shouldBeInvalid('1:23');
shouldBeInvalid('011:11');
shouldBeInvalid('ab:11');
shouldBeInvalid('-1:11');
shouldBeInvalid('24:11');
shouldBeInvalid('11');
shouldBeInvalid('11-');
shouldBeInvalid('11:-2');
shouldBeInvalid('11:60');
shouldBeInvalid('11:2b');
shouldBeInvalid('11:ab');
shouldBeInvalid('11:034');
shouldBeInvalid('23:45:');
shouldBeInvalid('23:45:6');
shouldBeInvalid('23:45:-1');
shouldBeInvalid('23:45:70');
shouldBeInvalid('23:45:zz');
shouldBeInvalid('23:45:06.');
shouldBeInvalid('23:45:06.abc');
shouldBeInvalid('23:45:06.789abc');

// Disabled
shouldBeValid('invalid', true);

var successfullyParsed = true;
</script>
<script src="../../js/resources/js-test-post.js"></script>
</body>
</html>
