LayoutTests/fast/forms/time/ValidityState-typeMismatch-time-expected.txt

This test aims to check for typeMismatch flag with type=time input fields

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS "" is a correct valid time string.
PASS "00:00" is a correct valid time string.
PASS "23:59" is a correct valid time string.
PASS "23:59:59" is a correct valid time string.
PASS "23:59:59.1" is a correct valid time string.
PASS "23:59:59.12" is a correct valid time string.
PASS "23:59:59.123" is a correct valid time string.
PASS "23:59:59.1234567890" is a correct valid time string.
PASS "00:00:00.0000000000" is a correct valid time string.
PASS " 00:00 " is a correct invalid time string.
PASS "1:23" is a correct invalid time string.
PASS "011:11" is a correct invalid time string.
PASS "ab:11" is a correct invalid time string.
PASS "-1:11" is a correct invalid time string.
PASS "24:11" is a correct invalid time string.
PASS "11" is a correct invalid time string.
PASS "11-" is a correct invalid time string.
PASS "11:-2" is a correct invalid time string.
PASS "11:60" is a correct invalid time string.
PASS "11:2b" is a correct invalid time string.
PASS "11:ab" is a correct invalid time string.
PASS "11:034" is a correct invalid time string.
PASS "23:45:" is a correct invalid time string.
PASS "23:45:6" is a correct invalid time string.
PASS "23:45:-1" is a correct invalid time string.
PASS "23:45:70" is a correct invalid time string.
PASS "23:45:zz" is a correct invalid time string.
PASS "23:45:06." is a correct invalid time string.
PASS "23:45:06.abc" is a correct invalid time string.
PASS "23:45:06.789abc" is a correct invalid time string.
PASS "invalid" is a correct valid time string when disabled.
PASS successfullyParsed is true

TEST COMPLETE
